Entity classes added.
Made some tests to check if the DAOs are working. Test database is set to H2, main DB stays on mysql.

+++ src/main/java/cz/cvut/fit/corsys/bl/service/DepartmentService.java
@@ -8,5 +8,5 @@ public interface DepartmentService {
 
     List<Department> findAllDepartments();
 
-    Department getDeparment(Long id);
+    Department getDeparment(Integer id);
 }
