+++ src/main/java/cz/cvut/fit/corsys/dl/dao/DepartmentDao.java
@@ -3,6 +3,6 @@ package cz.cvut.fit.corsys.dl.dao;
 import cz.cvut.fit.corsys.dl.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentDao extends JpaRepository<Department, Long> {
+public interface DepartmentDao extends JpaRepository<Department, Integer> {
 
 }
